feat: add support for multi-selection of files in the BranchView
The FileController in the BranchView now supports multi-selection of files. When a file is already selected and the user holds shift and clicks on another file, all files between the first and second selected file will be selected. If the index of the second file is smaller than the first, files upward from the first to the second selected file will be selected. If the index of the second file is bigger than the first, files downward from the first to the second selected file will be selected.