chore: add logging to FileController and UserController
The logging functionality is added to the FileController and UserController classes to improve debugging and monitoring. The MyLogger class is imported from the utility package. The logging level is set to FINE for both classes. Logging statements are added to various methods in both classes to log important events and information.